+++ modules/uhsg_eduviewer/src/Eduviewer.php
@@ -7,7 +7,7 @@ use Drupal\uhsg_active_degree_programme\ActiveDegreeProgrammeService;
 
 class Eduviewer {
 
-  const INVALID_DEGREE_PROGRAMME_CODES = ['KH20_001', 'MH30_003'];
+  const INVALID_DEGREE_PROGRAMME_CODES = ['KH20_001', 'MH30_001', 'MH30_003'];
 
   /** @var \Drupal\uhsg_active_degree_programme\ActiveDegreeProgrammeService*/
   private $activeDegreeProgrammeService;
@@ -36,7 +36,7 @@ class Eduviewer {
 
     if ($this->isValidDegreeProgrammeCode($activeDegreeProgrammeCode)) {
       $language = $this->languageManager->getCurrentLanguage()->getId();
-      $markup = "<div id=\"eduviewer-root\" degree-program-id=\"$activeDegreeProgrammeCode\" lang=\"$language\"></div>";
+      $markup = "<div id=\"eduviewer-root\" degree-program-id=\"$activeDegreeProgrammeCode\" lang=\"$language\" disable-global-style=\"true\"></div>";
     }
 
     return $markup;