+++ themes/uhsg_theme/js/otherEducationProviderSwitcher.js
@@ -2,6 +2,7 @@
   'use strict';
   Drupal.behaviors.otherEducationProviderSwitcher = {
     attach: function (context, settings) {
+      var triggerToggle = this.triggerToggle;
       var otherEducationProviderSwitcher = '.other-education-provider-switcher';
       var container = $('.other-education-provider-switcher');
       var header = $('.other-education-provider-switcher__header', otherEducationProviderSwitcher);
@@ -10,15 +11,17 @@
       var toggleIconClosed = 'icon--caret-down';
       var toggleIconOpen = 'icon--caret-up';
 
-      // toggle collapsed when clicking header
-      header.once().on('click', function () {
-        container.toggleClass(toggleClass);
-        $('body').toggleClass('no-scroll-mobile');
-        toggle.toggleClass(toggleIconClosed);
-        toggle.toggleClass(toggleIconOpen);
+      // Toggle collapsed when click or keypress on header
+      header.once().on({
+        click: function (event) {
+          triggerToggle(event, container, toggleClass, toggle, toggleIconClosed, toggleIconOpen);
+        },
+        keypress: function (event) {
+          triggerToggle(event, container, toggleClass, toggle, toggleIconClosed, toggleIconOpen);
+        }
       });
 
-      // close when clicking outside
+      // Close when clicking outside
       $(document).once().on('click', function (e) {
         var clickedOutside = $(e.target).parents(otherEducationProviderSwitcher).length === 0;
         if (container.hasClass(toggleClass) && clickedOutside) {
@@ -28,6 +31,14 @@
           toggle.toggleClass(toggleIconOpen);
         }
       });
+    },
+
+    triggerToggle: function (event, container, toggleClass, toggle, toggleIconClosed, toggleIconOpen) {
+      event.preventDefault();
+      container.toggleClass(toggleClass);
+      $('body').toggleClass('no-scroll-mobile');
+      toggle.toggleClass(toggleIconClosed);
+      toggle.toggleClass(toggleIconOpen);
     }
   };
 }(jQuery));