+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## 1.xx-dev
 Release date: ??.??.????
 
-* Improve bulletin arhive page (HUB-406)
+* Translate "User group" (HUB-393)
+* Include only the titles and the link in news email (HUB-425)
+* Exclude study track code from the Eduviewer embed (HUB-433)
+* Improve bulletin archive page (HUB-406)
 
 
 ## 1.30