+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release date: ??.??.????
 * Translate "User group" (HUB-393)
 * Include only the titles and the link in news email (HUB-425)
 * Exclude study track code from the Eduviewer embed (HUB-433)
+* Show summary field always when maintaining bulletins (HUB-357)
 
 
 ## 1.30