+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Release date: ??.??.????
 * Improve bulletin archive page (HUB-406)
 * Show summary field always when maintaining bulletins (HUB-357)
 * Fixed layout issues on pages that embed Eduviewer (HUB-428)
+* Limit summary field to 190 characters (HUB-358)
 * Add IE11 support to my searches (HUB-403)