+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Release date: ??.??.????
 * New taxonomy and related functionality: Other education provider (HUB-427)
 * Add keyboard navigation support to degree programme switcher (HUB-415)
 * Technical updates to npm dependencies (HUB-432)
+* Accessibility improvements (HUB-415, HUB-416, HUB-420)
 
 
 ## 1.30