+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Release date: ??.??.????
 * Add keyboard navigation support to degree programme switcher (HUB-415)
 * Technical updates to npm dependencies (HUB-432)
 * Accessibility improvements (HUB-415, HUB-416, HUB-420)
+* Update Drupal core (HUB-441)
 
 
 ## 1.30